requirements.txt
Django==2.2
nltk==3.5
numpy==1.19.1
pandas==1.1.1
django_filter==2.4.0
requests==2.24.0
djangorestframework==3.12.2
networkx==2.5
scikit_learn==0.23.2
